feat : 24.7 评论分页查询接口（2）- redis 缓存

+++ xiaohashu-comment/xiaohashu-comment-biz/src/main/java/com/quanxiaoha/xiaohashu/comment/biz/domain/dataobject/CommentDO.java
@@ -45,7 +45,7 @@ public class CommentDO {
 
     private Long childCommentTotal;
 
-    private BigDecimal heat;
+    private Double heat;
 
     private Long firstReplyCommentId;
